[LEMS-3000/ax-empty-helper] rename files to match function name
