+++ packages/perseus-score/src/has-empty-diner-widgets.ts
